Refine the neon depthwise convolution code(separate the Function and kernel).
